Added files via upload
activity_datatofetch.xml -  used to display the user interface for the Main or opening screen with "search movie" option and "view my list" option. 
data_movie.xml - used to display the user interface for the information fetched by the server of the queried movie and with button to "add movie to watch list".
fav_movie.xml -  used to display the user interface for the movies list of the user which has been create by the user by adding movie to his watch list.